pyproject.toml
[project]
name = "confluence-dump"
version = "0.1.0"
description = "Export Confluence pages to Markdown for LLM context"
readme = "README.md"
authors = [
    { name = "cheerchen", email = "cheerchen37@gmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.32.0",
    "click>=8.1.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=1.2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
confluence-dump = "confluence_dump.main:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
